added navigation to project
